userservice/requirements.txt
Django>=4.0,<5.0
Pillow>=9.0,<10.0
psycopg2-binary
djangorestframework>=3.14,<4.0
locust==3.0.0